+++ app/layout.jsx
@@ -3,12 +3,7 @@ import {Inter} from 'next/font/google'
 import Menu from "@/components/menu/Menu";
 import React from "react";
 import AppBanner from "@/components/menu/jabamaAppNotif/AppBanner";
-import HeaderPop from "@/components/header/headerPop/HeaderPop";
-import HeaderLanding from "@/components/header/HeaderLanding";
 import Footer from "@/components/footer/Footer";
-import HeaderPopDesktop from "@/components/header/headerPop/HeaderPopDesktop";
-import HeaderDesktop from "@/components/header/HeaderDesktop";
-import { Providers } from '@/redux/provider/provider';
 
 
 const inter = Inter({subsets: ['latin']})
@@ -24,9 +19,8 @@ export default function RootLayout({children}) {
     return (
         <html lang="en">
         <body className={inter.className}>
-        <Providers>
-            {children}
-        </Providers>
+
+        {children}
 
         <AppBanner display="lg:hidden"/>
         <Menu display="lg:hidden"/>
